requirements.txt
asgiref==3.5.2
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
flake8==4.0.1
isort==5.10.1
mccabe==0.6.1
openpyxl==3.0.10
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
